Split `createVault()` and reorganize class files
