+++ munkey/services/vault/index.ts
@@ -1,5 +1,9 @@
-import VaultService from "./VaultDatabase";
+import VaultDatabase from "./VaultDatabase";
+import VaultService from "./VaultService";
 
 export default VaultService;
 export * from "./VaultDatabase";
 export * from "./constructor";
+export {
+    VaultDatabase,
+};
